requirements.txt
sentence-transformers
matplotlib
numpy
scipy
midiutil
pillow
numba
